Move findLoginIDInPreviousInput
